sim & vis not done
