+++ .gitignore
@@ -236,3 +236,7 @@ modules.order
 Module.symvers
 Mkfile.old
 dkms.conf
+
+# other 
+
+data/*
