+++ src/data_process/README.md
@@ -1,0 +1,16 @@
+# Data for [Covid-19-Vaccine-Simultator](https://github.com/JJShen2000/Covid-19-Vaccine-Simultator)
+
+## Usage
+
+```
+python3 data_process.py
+```
+
+## Schema
+
+### town_population.csv
+Population of each census tracts in different time period in a day.
+
+| row1 | row2 | row3 | row4 | row5 | row6 |
+| --- | --- | --- | --- | --- | --- |
+| City | Census tract(town) | Workday night time population | Workday day time population | Weekend night time population | Weekend day time population
