+++ src/data_process/data_process.sh
@@ -1,0 +1,1 @@
+python3 data_process.py
