+++ src/data_process/raw_data/city_to_city_commute_raw.csv
@@ -1,0 +1,21 @@
+,新北市,臺北市,基隆市,新竹市,宜蘭縣,桃園市,新竹縣,臺中市,苗栗縣,彰化縣,南投縣,雲林縣,臺南市,高雄市,嘉義市,嘉義縣,屏東縣,澎湖縣,臺東縣,花蓮縣,其他地區
+新北市,1 300 214, 540 864, 8 704, 6 515, 1 521, 67 987, 4 356, 2 829, 1 283,  410,  274,  441,  844, 1 078,  235,  303,  225,  2,  96,  507, 1 643
+臺北市, 152 877,1 047 589, 5 015, 5 355, 1 208, 27 004, 3 028, 1 755,  966,  311,  222,  178,  781,  802,  84,  160,  100,  5,  120,  404, 1 342
+基隆市, 26 284, 38 267, 106 509,  505,  277, 2 040,  237,  302,  60,  32,  10,  18,  70,  94,  4,  6,  29,-,  3,  43,  246
+新竹市, 1 311, 3 201,  84, 183 717,-, 3 301, 21 627,  643, 3 877,  39,  28,  9,  82,  62,  15,  23,  6,-,-,-,  140
+宜蘭縣, 2 116, 2 853,  267,  18, 189 642,  270,  19,  183,  2,-,  2,-,-,  10,-,-,  13,-,  7,  700,  29
+桃園市, 47 537, 47 153,  701, 8 939,  206, 875 774, 14 169, 1 163, 1 099,  282,  99,  100,  379,  471,  19,  78,  91,  19,  1,  83,  504
+新竹縣, 1 501, 2 499,  27, 58 018,  23, 8 331, 153 313,  394, 2 790,  37,  15,  49,  85,  114,  2,  14,  16,-,  1,  20,  35
+臺中市, 1 695, 2 561,  59, 2 432,  54, 1 431, 1 157,1 227 693, 5 150, 17 923, 8 181, 1 411,  770,  717,  318,  543,  118,  7,  49,  68,  530
+苗栗縣,  733,  941,  29, 21 520,  12, 1 361, 6 274, 10 284, 194 994,  171,  63,  53,  96,  99,  3,  31,-,-,-,-,  33
+彰化縣,  380,  416,-,  312,-,  253,  143, 29 601,  172, 528 238, 5 038, 4 141,  332,  238,  158,  264,  37,-,  2,-,  53
+南投縣,  266,  194,-,  68,  4,  65,  69, 14 133,  102, 3 530, 195 494, 1 403,  67,  32,  80,  115,  10,-,  4,  28,  38
+雲林縣,  198,  120,  4,  46,-,  70,  49, 1 161,  18, 1 324,  491, 273 341,  404,  112, 1 621, 3 481,  20,-,  12,-,-
+臺南市,  790, 1 167,  10,  327,-,  346,  129, 1 229,  71,  372,  94,  823, 857 880, 16 726, 2 146, 3 085,  433,  16,  56,  48,  148
+高雄市, 1 119, 1 359,  63,  330,  30,  613,  295, 1 389,  160,  249,  75,  608, 31 403,1 261 472,  258,  757, 15 941,  78,  125,  81,  437
+嘉義市,  56,  32,-,  25,-,  26,  16,  200,  29,  73,  71, 2 044, 2 005,  118, 93 348, 21 490,  5,-,-,-,-
+嘉義縣,  124,  23,-,  5,  8,  48,  2,  316,  9,  209,  126, 4 729, 4 488,  186, 24 891, 185 833,  32,-,  15,  1,  34
+屏東縣,  127,  150,  4,  30,-,  103,  33,  129,  17,  28,  23,  58, 1 898, 38 618,  31,  47, 327 978,  18,  147,  77,  214
+澎湖縣,-,  21,-,-,-,  4,-,-,-,-,-,-,  3,  24,-,  3,-, 36 739,-,-,  6
+臺東縣,  32,  35,-,  17,-,  5,  4,  36,-,-,-,-,  20,  84,  3,-,  173,-, 92 023,  219,  19
+花蓮縣,  216,  155,-,  37,  146,  101,  23,  43,  3,  3,  6,  2,  2,  12,  5,  1,  75,-,  335, 132 828,  13
