--- src/sim/1g1a.txt
@@ -1,103 +1,0 @@
-100 1 1
-1
-0.01
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
-0 1 0 0 1 0 0
