--- src/sim/Node.h
@@ -1,3 +1,0 @@
-#pragma once
-
-#define node_t int
