--- src/sim/main.cpp
@@ -1,9 +1,0 @@
-#include <iostream>
-
-using namespace std;
-
-int main(int argc, char* argv[]) {
-    cout << "hello world\n";
-
-    return 0;
-}
