--- visualization/README.md
@@ -1,16 +1,0 @@
-# README
-## Required packages
-- dash<br>
-  `pip install dash`
--  plotly<br>
-  `pip install plotly`
-  
-## Usage of visualization_toy_example_dash.py
-make sure `census_tracts.csv` is in the same folder.
-
-`python ./visualization_toy_example_dash.py` or `python .\visualization_toy_example_dash.py`
-
-The graph will be running on http://127.0.0.1:8050/.
-
-## Demo
-![image](https://user-images.githubusercontent.com/40858520/143589905-8545b0ac-6ea9-486d-8079-9d44cc327025.png)
